requirements.txt
streamlit>=1.32.2
PyPDF2>=3.0.1
langchain-text-splitters>=0.0.1
openai>=1.12.0
python-dotenv>=1.0.0
langchain-community>=0.0.29
sentence-transformers>=2.2.2
numpy>=1.26.0
protobuf>=3.20.0
tqdm>=4.66.0
transformers>=4.35.0
